+++ R/check_membership.R
@@ -13,18 +13,15 @@
 #' @return A list of tibbles where the key variable exists and its value is in the membership list
 #'
 #' @examples
-#' \dontrun{
 #' # Create sample data with student emails
-#' path <- file.path(find.package("tutorial.helpers"), "tests/testthat/fixtures/answers_html")
-#' 
+#' path <- system.file("extdata", "answers_html", package = "tutorial.helpers")
+#'
 #' tibble_list <- gather_submissions(path, "stop")
-#' 
-#' result <- check_membership(tibble_list, 
-#'                            key_var = "email", 
-#'                            membership = c("bluebird.jack.xu@gmail.com", 
+#'
+#' result <- check_membership(tibble_list,
+#'                            key_var = "email",
+#'                            membership = c("bluebird.jack.xu@gmail.com",
 #'                                           "hassan.alisoni007@gmail.com"))
-#' 
-#' }
 #' @export
 check_membership <- function(tibble_list, key_var, membership, verbose = FALSE) {